requirements.txt
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
colorama==0.4.6
google-api-core==2.19.0
google-auth==2.30.0
google-cloud-speech==2.26.0
googleapis-common-protos==1.63.1
grpcio==1.64.1
grpcio-status==1.62.2
idna==3.7
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
requests==2.32.3
rsa==4.9
six==1.16.0
urllib3==2.2.1
